Migrate message "from_owner" flag to "sender"
Since we will allow the admins to interfere (upon request), a simple
BooleanField will not suffice anymore to discriminate between finder and
owner messages. a proper "sender" field is required.

First, we add the new fields, then migrate the values over, and last,
remove the old "from_owner" flag. Update the view code to reflect this.